[BE] refactor: ControllerTest 추상화 및 build.gradle 의존성 변경 #598
- mocking test 의존성 추가
- ApiSnippets 패키지 app 모듈 내로 이동
