+++ backend/app/build.gradle
@@ -39,8 +39,9 @@ dependencies {
 
     // REST Docs & restdocs-api-spec
     testImplementation 'org.springframework.restdocs:spring-restdocs-restassured'
+    testImplementation 'org.springframework.restdocs:spring-restdocs-mockmvc'
     testImplementation "com.epages:restdocs-api-spec-restassured:0.18.2"
-    testImplementation "com.epages:restdocs-api-spec:0.18.2"
+    testImplementation "com.epages:restdocs-api-spec-mockmvc:0.18.2"
     testImplementation testFixtures(project(':core'))
 
     // Swagger UI
