+++ backend/app/src/test/java/com/onair/hearit/category/presentation/CategoryControllerTest.java
@@ -11,7 +11,7 @@ import com.epages.restdocs.apispec.ResourceSnippetParameters;
 import com.epages.restdocs.apispec.Schema;
 import com.onair.hearit.category.dto.CategoryResponse;
 import com.onair.hearit.common.dto.response.PagedResponse;
-import com.onair.hearit.core.docs.ApiDocSnippets;
+import com.onair.hearit.fixture.ApiDocSnippets;
 import com.onair.hearit.domain.Category;
 import com.onair.hearit.fixture.IntegrationTest;
 import io.restassured.RestAssured;
