+++ backend/app/src/test/java/com/onair/hearit/fixture/ControllerTest.java
@@ -1,0 +1,48 @@
+package com.onair.hearit.fixture;
+
+import com.onair.hearit.auth.config.ApiSecurityConfig;
+import com.onair.hearit.auth.infrastructure.jwt.JwtTokenProvider;
+import com.onair.hearit.log.exception.FilterExceptionLogger;
+import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.extension.ExtendWith;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.boot.test.autoconfigure.restdocs.AutoConfigureRestDocs;
+import org.springframework.context.annotation.Import;
+import org.springframework.restdocs.RestDocumentationContextProvider;
+import org.springframework.restdocs.RestDocumentationExtension;
+import org.springframework.restdocs.mockmvc.MockMvcRestDocumentation;
+import org.springframework.security.test.web.servlet.setup.SecurityMockMvcConfigurers;
+import org.springframework.test.context.bean.override.mockito.MockitoBean;
+import org.springframework.test.web.servlet.MockMvc;
+import org.springframework.test.web.servlet.setup.MockMvcBuilders;
+import org.springframework.web.context.WebApplicationContext;
+
+@Import(ApiSecurityConfig.class)
+@ExtendWith(RestDocumentationExtension.class)
+@AutoConfigureRestDocs(outputDir = "build/generated-snippets")
+public abstract class ControllerTest {
+
+    @Autowired
+    protected MockMvc mockMvc;
+
+    @Autowired
+    private WebApplicationContext webApplicationContext;
+
+    @Autowired
+    private RestDocumentationContextProvider restDocumentationContextProvider;
+
+    @MockitoBean
+    protected JwtTokenProvider jwtTokenProvider;
+
+    @MockitoBean
+    protected FilterExceptionLogger filterExceptionLogger;
+
+    @BeforeEach
+    public void setUpMockMvc() {
+        this.mockMvc = MockMvcBuilders
+                .webAppContextSetup(webApplicationContext)
+                .apply(SecurityMockMvcConfigurers.springSecurity())
+                .apply(MockMvcRestDocumentation.documentationConfiguration(restDocumentationContextProvider))
+                .build();
+    }
+}
